backend/requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Message streaming
kafka-python==2.0.2

# Database and ORM
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
alembic==1.13.2

# Data processing
pandas==2.2.2

# Async and utilities
python-dotenv==1.0.1
requests==2.32.3

# Optional: logging, linting, testing
loguru==0.7.2
pytest==8.3.1
